Moved view tests to parallel
