+++ Tests/UnitTestsParallelizable/View/InitTests.cs
