+++ Tests/UnitTestsParallelizable/View/Navigation/HasFocusChangeEventTests.cs
