+++ Tests/UnitTestsParallelizable/View/Navigation/VisibleTests.cs
